refactor(common): move pre-team-preprocessing subpipeline from analytics to common
Now used by analytics, will be used by heatmaps + ai + clientwarnings
once they pick up the joined-shape pipeline. Path:
common/subpipelines/pre-team-preprocessing.ts.
